Deduplicate feather_m4 and metro_m4 blinky_rtc example
