+++ Cargo.toml
@@ -7,6 +7,8 @@ members = [
 	"pac/*",
 	"boards/*",
 ]
+# Fragments of example files, referenced by `include!()` from BSP examples
+exclude = ["boards/examples"]
 
 [profile.dev]
 debug = true
