+++ boards/feather_m4/examples/blinky_rtic.rs
@@ -1,77 +1,5 @@
-//! Blink an led using an RTIC software task and the RTC-based monotonic.
-
 #![no_std]
 #![no_main]
 
-use bsp::{hal, Pins, RedLed};
 use feather_m4 as bsp;
-#[cfg(not(feature = "use_semihosting"))]
-use panic_halt as _;
-#[cfg(feature = "use_semihosting")]
-use panic_semihosting as _;
-
-use hal::clock::v2::{clock_system_at_reset, rtcosc::RtcOsc};
-use hal::prelude::*;
-use hal::rtc::rtic::rtc_clock;
-use rtic::app;
-
-hal::rtc_monotonic!(Mono, rtc_clock::Clock1k);
-
-#[app(device = bsp::pac, dispatchers = [EVSYS_0])]
-mod app {
-    use super::*;
-
-    #[local]
-    struct Resources {}
-
-    #[shared]
-    struct Shared {
-        // The LED could be a local resource, since it is only used in one task
-        // But we want to showcase shared resources and locking
-        red_led: RedLed,
-    }
-
-    #[init]
-    fn init(cx: init::Context) -> (Shared, Resources) {
-        let device = cx.device;
-        let mut core: rtic::export::Peripherals = cx.core;
-
-        // Use v2 of the clocks API so that we can set the RTC clock source
-        let (_, clocks, tokens) =
-            clock_system_at_reset(device.oscctrl, device.osc32kctrl, device.gclk, device.mclk);
-
-        // Enable the RTC clock with the internal 1 kHz source.
-        // Note that currently the proof of this (the `RtcOsc` instance) is not
-        // required to start the monotonic.
-        let _ = RtcOsc::enable(tokens.rtcosc, clocks.osculp.osculp1k);
-
-        // Start the monotonic
-        Mono::start(device.rtc);
-
-        let pins = Pins::new(device.port);
-
-        // We can use the RTC in standby for maximum power savings
-        core.SCB.set_sleepdeep();
-
-        blink_led::spawn().ok().unwrap();
-
-        (
-            Shared {
-                red_led: pins.d13.into_push_pull_output(),
-            },
-            Resources {},
-        )
-    }
-
-    /// This function is spawned and never returns.
-    #[task(priority = 1, shared=[red_led])]
-    async fn blink_led(mut cx: blink_led::Context) {
-        loop {
-            // If the LED were a local resource, the lock would not be necessary
-            cx.shared.red_led.lock(|led| {
-                led.toggle().unwrap();
-            });
-            Mono::delay(400u64.millis()).await;
-        }
-    }
-}
+include!("../../examples/m4-blinky_rtic.rs");
